server_identity_service/requirements.txt
SQLAlchemy==1.3.12
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-RESTful==0.3.7
pyOpenSSL==19.1.0
PyJWT==1.7.1
SQLAlchemy-Utils==0.36.1